refactor: Zustand 스토어 구조 정리(#70)
- useUserStore 제거 (authStore로 통합)
- stores → store 폴더 통일
- webcamStore 이동 및 export 방식 통일
